Removed unused, and standardized the remaining names for all Kontract meta-schema and validatable value instances
