+++ kontracts/src/main/kotlin/io/amichne/kontracts/dsl/JsonSchemaBuilderDsl.kt
@@ -1,0 +1,4 @@
+package io.amichne.kontracts.dsl
+
+@DslMarker
+annotation class JsonSchemaBuilderDsl
