+++ kontracts/src/main/kotlin/io/amichne/kontracts/dsl/StringSchemaBuilder.kt
@@ -1,0 +1,31 @@
+package io.amichne.kontracts.dsl
+
+import io.amichne.kontracts.schema.StringSchema
+
+@JsonSchemaBuilderDsl
+open class StringSchemaBuilder : JsonSchemaBuilder {
+    var title: String? = null
+    var description: String? = null
+    var default: Any? = null
+    var nullable: Boolean = false
+    var example: Any? = null
+    var deprecated: Boolean = false
+    var minLength: Int? = null
+    var maxLength: Int? = null
+    var pattern: String? = null
+    var format: String? = null
+    var enum: List<String>? = null
+    fun build() = StringSchema(
+        title,
+        description,
+        default,
+        nullable,
+        example,
+        deprecated,
+        minLength,
+        maxLength,
+        pattern,
+        format,
+        enum
+    )
+}
