+++ kontracts/src/main/kotlin/io/amichne/kontracts/schema/ValidationResult.kt
@@ -1,5 +1,7 @@
 package io.amichne.kontracts.schema
 
+import kotlin.reflect.KClass
+
 /**
  * Result of schema validation.
  */
@@ -13,5 +15,10 @@ sealed class ValidationResult {
     fun getErrorMessage(): String? = when (this) {
         is Invalid -> message
         is Valid -> null
+    }
+
+    companion object {
+        inline fun <reified T : JsonSchema> KClass<T>.typeCheck(schema: JsonSchema): ValidationResult =
+            if (isInstance(schema)) Valid else Invalid("Expected type ${T::class.simpleName}, but got $simpleName")
     }
 }
