+++ kontracts/src/main/kotlin/io/amichne/kontracts/value/JsonBoolean.kt
@@ -3,19 +3,13 @@ package io.amichne.kontracts.value
 import io.amichne.kontracts.schema.BooleanSchema
 import io.amichne.kontracts.schema.JsonSchema
 import io.amichne.kontracts.schema.ValidationResult
+import io.amichne.kontracts.schema.ValidationResult.Companion.typeCheck
 
 /**
  * JSON boolean value.
  */
 data class JsonBoolean(val value: Boolean) : JsonValue() {
-    override fun validate(schema: JsonSchema): ValidationResult {
-        return when (schema) {
-            is BooleanSchema -> ValidationResult.Valid
-            else -> ValidationResult.Invalid(
-                "Expected ${schema}, but got Boolean"
-            )
-        }
-    }
+    override fun validate(schema: JsonSchema): ValidationResult = BooleanSchema::class.typeCheck(schema)
 
     override fun toString() = value.toString()
 }
