+++ kontracts/src/main/kotlin/io/amichne/kontracts/value/JsonNull.kt
@@ -3,19 +3,13 @@ package io.amichne.kontracts.value
 import io.amichne.kontracts.schema.JsonSchema
 import io.amichne.kontracts.schema.NullSchema
 import io.amichne.kontracts.schema.ValidationResult
+import io.amichne.kontracts.schema.ValidationResult.Companion.typeCheck
 
 /**
  * JSON null value.
  */
 object JsonNull : JsonValue() {
-    override fun validate(schema: JsonSchema): ValidationResult {
-        return when (schema) {
-            is NullSchema -> ValidationResult.Valid
-            else -> ValidationResult.Invalid(
-                "Expected ${schema}, but got Null"
-            )
-        }
-    }
+    override fun validate(schema: JsonSchema): ValidationResult = NullSchema::class.typeCheck(schema)
 
     override fun toString() = "null"
 }
