+++ src/main/kotlin/io/amichne/konditional/core/features/KotlinClassFeature.kt
@@ -2,14 +2,14 @@ package io.amichne.konditional.core.features
 
 import io.amichne.konditional.context.Context
 import io.amichne.konditional.core.Namespace
-import io.amichne.konditional.core.types.DataClassEncodeable
+import io.amichne.konditional.core.types.KotlinClassEncodeable
 import io.amichne.konditional.core.types.KotlinEncodeable
 import io.amichne.kontracts.schema.ObjectSchema
 
 /**
  * Sealed interface for custom encodeable feature flags.
  *
- * DataClassFeature (despite its name) allows using any custom encodeable type as feature flag values,
+ * KotlinClassFeature (despite its name) allows using any custom encodeable type as feature flag values,
  * providing structured, type-safe configuration with full schema validation.
  *
  * **Note:** This interface name is retained for backwards compatibility but works with
@@ -43,27 +43,27 @@ import io.amichne.kontracts.schema.ObjectSchema
  * @param C The context type used for evaluation
  * @param M The namespace this feature belongs to
  */
-sealed interface DataClassFeature<T : KotlinEncodeable<ObjectSchema>, C : Context, M : Namespace> :
-    Feature<DataClassEncodeable<T>, T, C, M> {
+sealed interface KotlinClassFeature<T : KotlinEncodeable<ObjectSchema>, C : Context, M : Namespace> :
+    Feature<KotlinClassEncodeable<T>, T, C, M> {
 
     companion object {
         /**
-         * Factory function for creating DataClassFeature instances.
+         * Factory function for creating KotlinClassFeature instances.
          *
          * @param key The feature key (usually the property name)
          * @param module The namespace this feature belongs to
-         * @return A DataClassFeature instance
+         * @return A KotlinClassFeature instance
          */
         operator fun <T : KotlinEncodeable<ObjectSchema>, C : Context, M : Namespace> invoke(
             key: String,
             module: M,
-        ): DataClassFeature<T, C, M> =
-            DataClassFeatureImpl(key, module)
+        ): KotlinClassFeature<T, C, M> =
+            KotlinClassFeatureImpl(key, module)
 
         @PublishedApi
-        internal data class DataClassFeatureImpl<T : KotlinEncodeable<ObjectSchema>, C : Context, M : Namespace>(
+        internal data class KotlinClassFeatureImpl<T : KotlinEncodeable<ObjectSchema>, C : Context, M : Namespace>(
             override val key: String,
             override val namespace: M,
-        ) : DataClassFeature<T, C, M>
+        ) : KotlinClassFeature<T, C, M>
     }
 }
