+++ src/main/kotlin/io/amichne/konditional/core/types/KotlinClassExtensions.kt
