+++ src/test/kotlin/io/amichne/konditional/adversarial/AdversarialConfigTest.kt
@@ -1,5 +1,6 @@
 package io.amichne.konditional.adversarial
 
+import io.amichne.konditional.api.evaluate
 import io.amichne.konditional.context.AppLocale
 import io.amichne.konditional.context.Context
 import io.amichne.konditional.context.Platform
@@ -7,7 +8,6 @@ import io.amichne.konditional.context.Rampup
 import io.amichne.konditional.context.Version
 import io.amichne.konditional.core.Namespace
 import io.amichne.konditional.core.features.FeatureContainer
-import io.amichne.konditional.core.features.evaluate
 import io.amichne.konditional.core.id.StableId
 import io.amichne.konditional.fixtures.core.TestNamespace
 import io.amichne.konditional.fixtures.core.test
