+++ src/test/kotlin/io/amichne/konditional/core/EnumFeatureTest.kt
@@ -6,7 +6,7 @@ import io.amichne.konditional.context.Platform
 import io.amichne.konditional.context.Version
 import io.amichne.konditional.core.features.EnumFeature
 import io.amichne.konditional.core.features.FeatureContainer
-import io.amichne.konditional.core.features.evaluate
+import io.amichne.konditional.api.evaluate
 import io.amichne.konditional.core.id.StableId
 import io.amichne.konditional.core.types.EncodableEvidence
 import io.amichne.konditional.core.types.EncodableValue
